Clean up and make module graph acyclic
